requirements.txt
streamlit
gspread
google-auth
openai>=1.1.0
python-docx
